requirements.txt
python-dotenv
groq
yfinance
duckduckgo-search
newspaper4k
lxml_html_clean
agno
sqlalchemy>=1.4
psycopg[binary]
pgvector
pypdf
udocker
sentence-transformers
mcp
